Widen AccountId from int to long end-to-end
The diagnostic log on the SpawnPlayerPacket hook showed the hook
firing cleanly with verified per-player ulong AccountIds direct from
the network packet — values like 4916870491923041932,
14140241474363346572, 4916873569267109516, etc. Every row distinct,
exactly what alt-linkage needs.

The reason it still looked broken: the entire pipeline had been
storing AccountId as int (signed 32-bit). Casting a ulong to int32
discards the upper 32 bits, which are non-zero and account-relevant
in FFXIV's id allocation. So thousands of distinct real ulong
AccountIds collapsed into the same int32 buckets, and what looked
like leakage / fake alts was actually truncation collisions. The
Brynhildr-account leaking onto a Midgardsormr stranger was two
distinct full ulongs whose lower 32 bits happened to match.

bc->AccountId in the Character struct was reliable the whole time.
The spawn-packet hook is reliable. Object-table reads are reliable.
The cast to int32 was the only thing destroying uniqueness.

This commit widens the field to long? everywhere it lives —
Player.AccountId entity column (EF migration AlterColumn int→bigint),
all PostPlayerRequest / PlayerDto / PlayerDetailed / PlayerSearchDto
mirrors on both sides. The plugin's scan sites use unchecked
((long)bc->AccountId) so the bit pattern of the original ulong is
preserved across the signed-vs-unsigned cast (we never need to
arithmetic on AccountId, only equality-match for grouping, and bit
identity is sufficient).

AlphaScope's local-SQLite-cache used ulong throughout, which is why
the test that worked there fell apart on Memoria's centralized
schema — the int? column type was an undocumented downgrade.

+++ Memoria/API/Models/Requests/Player/PostPlayerRequest.cs
@@ -18,7 +18,7 @@ namespace Memoria.API.Models.Requests.Player
         [JsonProperty("3")]
         public ushort? HomeWorldId { get; set; }
         [JsonProperty("4")]
-        public int? AccountId { get; set; }
+        public long? AccountId { get; set; }
         [JsonProperty("5")]
         public short? TerritoryId { get; set; }
         [JsonProperty("6")]
